% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{get_db_connection}
\alias{get_db_connection}
\title{Get SQLite database connection}
\usage{
get_db_connection()
}
\value{
connection
}
\description{
Get SQLite database connection
}
\examples{
get_db_connection()
}
